Making an attempt to move to LiveComponent
I realized that the functionality in the PlayController can be done with
Plugs, meaning we can probably get rid of it entirely. Gonna be doing
that shortly.

I made an attempt at making another LiveView (PlayLiveView) that would
take a LiveComponent as the view implementation rather than a custom
behaviour like it's doing now. I'm still kinda struggling with that so
just gonna drop it for now and work on removing the controller.