Minor cleanup, and splitting the GameServer tests.
Was hoping this would make them run a little faster, but it turns out
Mock can't be used in async tests. Keeping it, though, since these are
much easier to read.

Also rearranged some functions in Minesweeper.GameState and added more
to the View guide since I forgot to mention a big feature.